+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 7.7.0 Fixes
 
+- `[Datagrid]` Added support to disable column buttons. ([1590](https://github.com/infor-design/enterprise/issues/1590))
 - `[Popupmenu]` Expose is-selectable, is-multiselectable, and multi-selectable-section as input properties. ([#907](https://github.com/infor-design/enterprise-ng/issues/907)) `CL`
 
 ## v7.6.0